pyproject.toml
[tool.poetry]
name = "seefood"
version = "0.1.0"
description = ""
authors = ["Mike Czech <mike.czech@ottogroup.com>"]

[tool.poetry.dependencies]
python = "^3.7"
nb_black = "^1.0.7"
jupyterlab = "^2.0.1"
tqdm = "^4.43.0"
pandas = "^1.0.3"
altair = "^4.0.1"
scikit-learn = "^0.22.2"
torch = "^1.4.0"
torchvision = "^0.5.0"
pillow = "^7.0.0"
wordcloud = "^1.6.0"
matplotlib = "^3.2.1"
black = "^19.10b0"
google-cloud-bigquery = {extras = ["pandas"], version = "^1.24.0"}
tensorboard = "^2.1.1"
ipywidgets = "^7.5.1"
line_profiler = "^3.0.2"
lmdb = "^0.98"
pympler = "^0.8"
pandas_gbq = "^0.13.1"
flask = "^1.1.2"
umap-learn = "^0.4.1"
bokeh = "^2.0.1"
datashader = "^0.10.0"
colorcet = "^2.0.2"
holoviews = "^1.13.2"
twisted = "^20.3.0"
skorch = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
